WIP: Rewrote evaluation system, check README
